move ItemCategory methods to their own Utils class
