+++ RuntimeIcons/src/RuntimeIcons.cs
@@ -81,7 +81,7 @@ public class RuntimeIcons : BaseUnityPlugin
 
             Log.LogInfo("Patching Methods");
 
-            CategorizeItemPatch.Init();
+            StartOfRoundPatch.Init();
             Harmony.PatchAll();
 
             Log.LogInfo(NAME + " v" + VERSION + " Loaded!");
